Added some more iframes, optimized, and fixed bug
Added several more iframes to the portfolio

with more iframes, the filters were lagging because every change was assigning the src, reloading the page. Now the src is checked to see if it needs assigning

There was an issue where iframe src would stay loaded between page reloads and so hidden frames would, even though hidden, load and message the main page with their height causing them to appear. Now there is a check to make sure the main page has fully loaded and we know which frames are active before received messages affect the frame height