Images in iframes now sent to parent frame to be expanded there
Because some descriptive text is shown when you click images, I want people to be able to access that info easily.

Before, an iframe would expand the image inside the iframe, but this forces the user to scroll all the way to the top of the iframe to view the image.

To resolve this, if the image is in an iframe, the page now detects this and sends the image info as a message to the main page to get expanded there instead